added library initialization with config, AssetPair class creation
