+++ .gitignore
@@ -4,6 +4,7 @@
 
 dist/
 
+yarn-error.log
 node_modules
 **/*.js
 **/*.js.map
